+++ src/org/elixir_lang/psi/Unquote.kt
@@ -49,16 +49,19 @@ object Unquote {
                     .filter(ResolveResult::isValidResult)
                     .mapNotNull(ResolveResult::getElement)
                     .filterIsInstance<Call>()
+                    .filter { !resolveState.hasBeenVisited(it) }
                     .let { resolveds -> treeWalkUpUnquoted(resolveds, resolveState, keepProcessing) }
 
     private fun treeWalkUpUnquoted(unquotedList: List<Call>,
                                    resolveState: ResolveState,
                                    keepProcessing: (PsiElement, ResolveState) -> Boolean): Boolean =
             whileIn(unquotedList) { unquoted ->
+                val unquotedResolveState = resolveState.putVisitedElement(unquoted)
+
                 if (CallDefinitionClause.`is`(unquoted)) {
-                    Using.treeWalkUp(unquoted, null, resolveState, keepProcessing)
+                    Using.treeWalkUp(unquoted, null, unquotedResolveState, keepProcessing)
                 } else {
-                    treeWalkUpUnquotedVariable(unquoted, resolveState, keepProcessing)
+                    treeWalkUpUnquotedVariable(unquoted, unquotedResolveState, keepProcessing)
                     // a variable
 
                     true