+++ src/org/elixir_lang/psi/impl/PsiElementImpl.kt
@@ -20,8 +20,9 @@ import org.elixir_lang.psi.impl.call.maybeModularNameToModulars
 import org.elixir_lang.psi.operation.Match
 import org.elixir_lang.psi.operation.Pipe
 import org.elixir_lang.psi.scope.WhileIn.whileIn
+import org.elixir_lang.util.AccumulatorContinue
+import org.elixir_lang.util.foldWhile
 import org.jetbrains.annotations.Contract
-import java.util.*
 
 fun PsiElement.ancestorSequence() = generateSequence(this) { it.parent }
 fun PsiElement.document(): Document? = containingFile.viewProvider.let { viewProvider ->
@@ -111,9 +112,7 @@ tailrec fun PsiElement.selfOrEnclosingMacroCall(): Call? =
     }
 
 /**
- *
- * @param call
- * @return `null` if call is at top-level
+ * @return `null` if this element is at top-level
  */
 @Contract(pure = true)
 fun PsiElement.enclosingMacroCall(): Call? = parent.selfOrEnclosingMacroCall()
@@ -241,7 +240,7 @@ fun <R> PsiElement.childExpressionsFoldWhile(
         element: PsiElement, accumulator: R
     ) -> AccumulatorContinue<R>
 ): AccumulatorContinue<R> =
-    AccumulatorContinue.childExpressionsFoldWhile(this, forward, initial, folder)
+    childExpressions(forward).foldWhile(initial, folder)
 
 fun PsiElement.childExpressions(forward: Boolean = true): Sequence<PsiElement> {
     val seed = if (forward) {
@@ -263,11 +262,11 @@ fun PsiElement.isExpression(): Boolean =
     }
 
 @Contract(pure = true)
-fun PsiElement.siblingExpression(function: (PsiElement) -> PsiElement): PsiElement? {
-    var expression = this
+fun PsiElement.siblingExpression(function: (PsiElement) -> PsiElement?): PsiElement? {
+    var expression: PsiElement? = this
 
     do {
-        expression = function(expression)
+        expression = function(expression ?: return null)
     } while (expression is ElixirEndOfExpression ||
         expression is LeafPsiElement ||
         expression is PsiComment ||