+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/bare_head_preferred_usage.ex
@@ -1,0 +1,5 @@
+defmodule Prefix.BareHeadPreferredUsage do
+  alias Prefix.BareHeadPreferredDeclaration
+
+  BareHeadPreferredDeclaration.<caret>
+end