+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/no_bare_head_fallback_usage.ex
@@ -1,0 +1,5 @@
+defmodule Prefix.NoBareHeadFallbackUsage do
+  alias Prefix.NoBareHeadFallbackDeclaration
+
+  NoBareHeadFallbackDeclaration.<caret>
+end